fix: add hover to follow buttons
